break out kernel parts into separate files
